Correction to np type for MB_TYPE_DOUBLE.

+++ pymoab/types.pyx
@@ -67,7 +67,7 @@ MB_MAX_DATA_TYPE = moab.MB_MAX_DATA_TYPE
 _DTYPE_CONV = {
     MB_TYPE_OPAQUE   : np.int8,
     MB_TYPE_INTEGER  : np.int32,
-    MB_TYPE_DOUBLE   : np.float32,
+    MB_TYPE_DOUBLE   : np.float64,
     MB_TYPE_BIT      : 'yikes',
     MB_TYPE_HANDLE   : np.uint64,
     MB_MAX_DATA_TYPE : np.uint64
